changed function return type and added come comments

+++ main.cpp
@@ -1,22 +1,23 @@
 #include <iostream>
 
 // djb2 hash, popular for strings
-int djb2(std::string &str) {
-  int hash = 5381; // magic number
+unsigned long long djb2(std::string &str) {
+  unsigned long long hash = 5381; // magic number
   // for every char in str
   for (char c : str) {
-    hash = (hash << 5) + hash + c; // hash * 33 + c
+    hash = (hash << 5) + hash + c; // hash * 33 + c (the char)
   }
   return hash;
 }
 
 // fnv-1a hash, fast and general
-int fnv1a(std::string &str) {
+unsigned long long fnv1a(std::string &str) {
   unsigned long long hash = 14695981039346656037ULL; // magic number
   // for every char in str
   for (char c : str) {
     hash ^= static_cast<unsigned long long>(
-        c);                   // add random current bits in c (the char)
+        c); // add current bits from c (the char) and change the type to
+            // unsigned long long
     hash *= 1099511628211ULL; // multiply by FNV_prime
   }
   return hash;
